Move test app into spec/dummy
This isn't library code, it's test code. It doesn't belong in `lib`.
Common convention in engines is to have the dummy app in `spec/dummy`.

* Update namespace so it doesn't look like part of Clearance.
* Update reference from `Clearance::Testing::Application` to
  `Dummy::Application`.
* Move dummy rake tasks into the `dummy` rake namespace.
* Update bin/setup to use the new namespace.
* Update bin/setup to no longer clean old appraisals out.
* Update style in Rakefile and Dummy app
