+++ .travis.yml
@@ -4,15 +4,10 @@ language:
 rvm:
   - 1.9.3
   - 2.0.0
-  - 2.1.3
+  - 2.1.5
 
 install:
-  - "travis_retry bundle install"
-  - "bundle exec appraisal generate"
-  - "travis_retry bundle exec appraisal install"
-
-before_script:
-  - "bundle exec rake db:migrate"
+  - "travis_retry bin/setup"
 
 script:
   - "bundle exec appraisal rake"
