+++ bin/setup
@@ -5,8 +5,7 @@ set -e
 # Install required gems, including Appraisal, which helps us test against
 # multiple Rails versions
 bundle install
-bundle exec appraisal clean
 bundle exec appraisal install
 
 # Set up database for the application that Clearance tests against
-RAILS_ENV=test bundle exec rake db:reset
+RAILS_ENV=test bundle exec rake dummy:db:reset
